+++ src/agent/dietAgent.cc
@@ -10,6 +10,10 @@
 /****************************************************************************/
 /* $Id$
  * $Log$
+ * Revision 1.16  2004/07/05 14:56:13  rbolze
+ * correct bug on 64 bit plat-form, when parsing cfg file :
+ * remplace size_t by unsigned int for config options
+ *
  * Revision 1.15  2004/05/28 10:53:21  mcolin
  * change the endpoint option names for agents and servers
  *  endPointPort -> dietPort
@@ -155,7 +159,8 @@ main(int argc, char** argv)
 
   /* Get listening port & hostname */
 
-  size_t* port = (size_t*) 
+    // size_t --> unsigned int
+  unsigned int* port = (unsigned int*) 
     (Parsers::Results::getParamValue(Parsers::Results::DIETPORT));
   char* host = (char*)
     (Parsers::Results::getParamValue(Parsers::Results::DIETHOSTNAME));
@@ -197,13 +202,17 @@ main(int argc, char** argv)
 #if HAVE_LOGSERVICE
   /* Create the DietLogComponent */
   bool useLS;
-  size_t* ULSptr;
+    // size_t --> unsigned int
+  unsigned int* ULSptr;
   int outBufferSize;
-  size_t* OBSptr;
+    // size_t --> unsigned int
+  unsigned int* OBSptr;
   int flushTime;
-  size_t* FTptr;
+    // size_t --> unsigned int
+  unsigned int* FTptr;
 
-  ULSptr = (size_t*)Parsers::Results::getParamValue(
+    // size_t --> unsigned int
+  ULSptr = (unsigned int*)Parsers::Results::getParamValue(
               Parsers::Results::USELOGSERVICE);
   useLS = false;
   if (ULSptr == NULL) {
@@ -215,7 +224,8 @@ main(int argc, char** argv)
   }
 
   if (useLS) {
-    OBSptr = (size_t*)Parsers::Results::getParamValue(
+    // size_t --> unsigned int
+    OBSptr = (unsigned int*)Parsers::Results::getParamValue(
   	       Parsers::Results::LSOUTBUFFERSIZE);
     if (OBSptr != NULL) {
       outBufferSize = (int)(*OBSptr);
@@ -223,8 +233,8 @@ main(int argc, char** argv)
       outBufferSize = 0;
       WARNING("lsOutbuffersize not configured, using default");
     }
-
-    FTptr = (size_t*)Parsers::Results::getParamValue(
+    // size_t --> unsigned int
+    FTptr = (unsigned int*)Parsers::Results::getParamValue(
   	       Parsers::Results::LSFLUSHINTERVAL);
     if (FTptr != NULL) {
       flushTime = (int)(*FTptr);
